LibWeb: Move some static functions from CSS/Parser.cpp into a new file
